tags and updates are saved in real time and group delete also deletes all the tags inside the features
