+++ src/app/components/side-bar/tag-images/tag-images.component.ts
@@ -6,7 +6,8 @@ import { FormGroup } from '@angular/forms';
 import { Router } from '@angular/router';
 import { Subscription } from 'rxjs';
 import { GeoDataService } from 'src/app/services/geo-data.service';
-import { FeatureCollection } from 'geojson';
+import { Feature, FeatureCollection } from 'geojson';
+import { ProjectsService } from 'src/app/services/projects.service';
 
 @Component({
   selector: 'app-tag-images',
@@ -22,18 +23,22 @@ export class TagImagesComponent implements OnInit {
   private formGroup$: Subscription;
   private activeGroup$: Subscription;
   private activeFeatureId$: Subscription;
+  private groupList;
+  private selectedProject
   form: FormGroup;
   showSubitem: boolean = true;
   tagList: tags[] = this.formsService.getTags();
   newTag: tags[] = [];
   featureList: Array<any> = [];
   features: FeatureCollection;
+  tempGroup: Array<Feature>;
 
   constructor(
 	private groupsService: GroupsService,
 	private formsService: FormsService,
 	private dialog: MatDialog,
 	private router: Router,
+	private projectsService: ProjectsService,
 	private geoDataService: GeoDataService,) { }
 
   ngOnInit() {
@@ -47,6 +52,13 @@ export class TagImagesComponent implements OnInit {
 
 	this.formGroup$ = this.formsService.formGroup.subscribe((next) => {
 	  this.form = next;
+	});
+
+	this.groupsService.groups.subscribe((next) => {
+		this.groupList = next;
+	});
+	this.projectsService.activeProject.subscribe(next => {
+		this.selectedProject = next;
 	});
 
 	this.geoDataService.features.subscribe( (fc: FeatureCollection) => {
@@ -80,7 +92,89 @@ export class TagImagesComponent implements OnInit {
   openRenameOptionModal(template: TemplateRef<any>) {
 	this.dialog.open(template);
   }
+  
+  deleteTag(gName: string, tag: tags){
+	let data = this.tagList;
+	while(true){
+		const index = data.findIndex(item => item.groupName === gName && item.label === tag.label && item.type === tag.type);
+		// delete this.exampleNote[index];
+		if (index > -1) {
+		data.splice(index, 1);
+		}else{
+			break;
+		}
+	}
+	// deletes all the tags that doesn't have group names
+	// while(true){
+	// 	const index = data.findIndex(item => item.groupName === "");
+	// 	// delete this.exampleNote[index];
+	// 	if (index > -1) {
+	// 	data.splice(index, 1);
+	// 	}else{
+	// 		break;
+	// 	}
+	// }
 
+
+	this.tagList = data;
+
+	let icon:string
+	let payload
+	this.groupList.forEach(group => {
+		if (group.name == this.activeGroup) {
+			this.tempGroup = group.features;
+				icon = group.icon
+	
+				//Creates a temporary group with a copy of the current groups info
+				let tempGroup = [{
+					name: group.name,
+					color: group.color,
+					icon: group.icon
+				}]
+				
+				//And adds the temp group to a payload along with the necessary style infromation
+				payload = {
+					group: tempGroup,
+					style: {
+						faIcon: icon,
+						color: '#00C8FF'
+					},
+					tag: []
+				}
+				// console.log(payload.tag)
+			
+		}
+		});
+
+		// console.log(this.tempGroup)
+		for (let feat of this.tempGroup){
+
+			if(feat.properties.tag != undefined || feat.properties.tag != []){
+				// feat.properties.tag.forEach(tag => {
+				//   this.newTag.push(tag)
+				// });
+				// console.log(feat.properties)
+				// console.log(this.activeGroup)
+				feat.properties.group.forEach(group => {
+					if(group.name != this.activeGroup){
+						let tempGroup = {
+							name: group.name,
+							color: group.color,
+							icon: group.icon
+						}
+					payload.group.push(tempGroup)}
+				});
+			} 
+			
+			console.log(data)
+			payload.tag = data;
+
+			this.geoDataService.updateFeatureProperty(this.selectedProject.id, Number(feat.id), payload)
+			payload.tag = []
+		}
+
+
+  }
   openOptionToggle(label: string) {
 	if (this.openOption[label]) {
 	  this.openOption[label] = false;
